requirements.txt
flask
PyPDF2
langchain-openai
langchain-community
faiss-cpu
tiktoken
python-dotenv
